Add Lego Power Functions tests
